Added function to Store Login Details into DB

+++ TestProject/members/models.py
@@ -3,5 +3,5 @@ from django.db import models
 # Create your models here.
 
 class userDetails(models.Model):
-    number = models.BigIntegerField(max_length=11)
+    number = models.IntegerField(max_length=11)
     password = models.CharField(max_length=16)
